Create MyCounter Library
The motivation to write this small extension was my need to have a counting process running to control the “time” in a simulation process. I needed something that would update a selected cell of a Calc’s sheet to represent time running.
This to be uses in LibreOffice and similar spreadsheets. To use in VBA requires changes in the communications to the Stopwatch Control and to the calling sheet and cell.