Create dlgCounter
This xmml file containd the code to create the dialof dlgCounter.xdl ro communicate back and forth with the .bas code. 
Xopy this xml in a text file. Save it with the name dlgCounter.xdl and import as a dialog in the same mibrary where the .bas code was imported
(LobreOffice Macro)