Create MyCounter Library (version 2.1.1)
This is the version submited to LibreOffice Extensions on 2024-01-24.
User needs to download and add the two dialogs dlgCounter1.xdl and dlgouter2.xdl. These are two option that provide two different button's behaviors. Users can design their own dialogs and, if needed, new methods to deal with other buttons's behaviors. These dialogs have eventdriven calls to methods in the library.
The extension instaler can be found in
https://extensions.libreoffice.org
Search for mycounter.